Bugfix and refactoring
Fixed a problem that quick outline was not displayed at the cursor
position even when the middle mouse button was clicked on the virtual
diagram.

On the main diagram, the problem that the table can not be retrieved
even if the table is selected by the quick outline has been corrected.